+++ tests/tinytest/test-endogenous-stats4.R
@@ -12,9 +12,9 @@ event_types <- c(1, 1, 2, 2, 1)
 # Statistics
 edgelist$type <- event_types
 reh <- remify::remify(edgelist, model = "tie", directed = FALSE)
-effects <- ~ inertia() + sp() + spUnique() + psABAB() + psABAY() +
+effects <- ~ inertia() + sp() + sp(unique = TRUE) + psABAB() + psABAY() +
   inertia(consider_type = TRUE) +
-  sp(consider_type = TRUE) + spUnique(consider_type = TRUE) +
+  sp(consider_type = TRUE) + sp(unique = TRUE, consider_type = TRUE) +
   psABAB(consider_type = TRUE) + psABAY(consider_type = TRUE)
 stats <- remstats(reh, tie_effects = effects)
 riskset <- attr(stats, "riskset")
@@ -50,7 +50,7 @@ sp <- rbind(
   c(0, 0, 1, 0, 0, 1),
   c(1, 1, 1, 1, 1, 1)
 )
-expect_equal(stats[, , "sp"], sp)
+expect_equal(stats[, , which(dimnames(stats)[[3]] == "sp")[1]], sp)
 
 # sp.type
 sp.type <- rbind(
@@ -60,7 +60,7 @@ sp.type <- rbind(
   c(0, 0, 1, 0, 0, 0),
   c(0, 0, 1, 0, 1, 0)
 )
-expect_equal(stats[, , "sp.type"], sp.type)
+expect_equal(stats[, , which(dimnames(stats)[[3]] == "sp.type")[1]], sp.type)
 
 # spUnique
 spUnique <- rbind(
@@ -70,7 +70,7 @@ spUnique <- rbind(
   c(0, 0, 1, 0, 0, 1),
   c(1, 1, 1, 1, 1, 1)
 )
-expect_equal(stats[, , "spUnique"], spUnique)
+expect_equal(stats[, , which(dimnames(stats)[[3]] == "sp")[2]], spUnique)
 
 # spUnique.type
 spUnique.type <- rbind(
@@ -80,7 +80,7 @@ spUnique.type <- rbind(
   c(0, 0, 1, 0, 0, 0),
   c(0, 0, 1, 0, 1, 0)
 )
-expect_equal(stats[, , "spUnique.type"], spUnique.type)
+expect_equal(stats[, , which(dimnames(stats)[[3]] == "sp.type")[2]], spUnique.type)
 
 # psABAB
 psABAB <- rbind(
@@ -124,10 +124,9 @@ expect_equal(stats[, , "psABAY.type"], psABAY.type)
 
 # test standardization
 std_effects <- ~
-  inertia(scaling = "std") + sp(scaling = "std") + spUnique(scaling = "std") + 
+  inertia(scaling = "std") + sp(scaling = "std") + 
   inertia(consider_type = TRUE, scaling = "std") + 
-  sp(consider_type = TRUE, scaling = "std") + 
-  spUnique(consider_type = TRUE, scaling = "std")
+  sp(consider_type = TRUE, scaling = "std") 
 std_stats <- remstats(reh, tie_effects = std_effects)
 
 sapply(2:dim(std_stats)[3], function(p) {